[Mosaic GPU] Implement GPU module unloading for Mosaic GPU custom calls.
We introduce reference counting in the global kernel cache. A kernel handle is created in the global cache the first time it is compiled. Subsequent calls to "compile" will return a shared pointer to the kernel handle. We keep references to the compiled kernels in the custom call state. We unload the module in the kernel handle destructor. I.e. when there is no more reference to a kernel handle, the module is unloaded.

We retain the old caching mechanism for the legacy custom call. GPU module are never unloaded when using the legacy custom call.

PiperOrigin-RevId: 837838718